requirements.txt
asgiref==3.8.1
attr==0.3.2
ConfigParser==7.1.0
contextlib2==21.6.0
cryptography==43.0.1
mysqlclient==2.2.4
Cython==3.0.11
dl==0.1.0
docutils==0.21.2
flask_sqlalchemy==3.1.1
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.1
#jnius==1.1.0
#keyring==25.4.1
#lockfile==0.12.2
marshmallow==3.20.1
mock==5.1.0
mysql_connector_repackaged==0.3.1
numpy==1.25.0
Pillow==10.0.0
#Pillow==10.4.0
protobuf==5.28.2
pyOpenSSL==24.2.1
python-dotenv==1.0.1
railroad==0.5.0
#Sphinx==8.0.2
thread==2.0.5
tornado==6.3.2
trove_classifiers==2024.9.12
#truststore==0.9.2
urllib3_secure_extra==0.1.0
watchdog==5.0.3
#xmlrpclib==1.0.1
